Support new export handling in Anki 2.1.55+
